application/requirements.txt
Fabric
Flask-WTF
WTForms
flask_pymongo
coverage
kubernetes
flask-login==0.4.1
Werkzeug==0.10.1